backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.25.1
pydantic[email]==2.5.0
cors==1.0.1
passlib==1.7.4
aiosmtplib==3.0.1
jinja2==3.1.2
requests==2.31.0
cryptography==41.0.7
